refactor(descargas): remove deprecated code informe rup
