+++ modules/descargas/informe-rup/elementos-rup/index.ts
@@ -1,19 +1,19 @@
-import { ObservacionesComponent } from './observaciones-component';
-import { MoleculaBaseComponent } from './molecula-base-component';
-import { ValorNumericoComponent } from './valor-numerico-component';
-import { ValorNumericoFactory } from './valor-numerico-factory';
-import { ValorFechaComponent } from './valor-fecha-component';
-import { SolicitudPrestacionDefaultComponent } from './solicitud-default-component';
-import { SelectPorRefsetComponent } from './select-por-refset-component';
-import { RegistrarMedicamentoDefaultComponent } from './registrar-medicamento-default-component';
-import { LugarDeNacimientoComponent } from './lugar-nacimiento-componenet';
-import { LactanciaComponent } from './lactancia-component';
-import { FormulaBaseComponent } from './formula-base-component';
-import { SeleccionBinariaComponent } from './seleccion-binaria-component';
-import { EvolucionProblemaDefaultComponent } from './evolucion-problema-default-component';
-import { CalculoDeBostonComponent } from './calculo-boston-component';
-import { AutocitadoComponent } from './autocitado-component';
-import { AdjuntarDocumentoComponent } from './adjuntar-documento-component';
+import { ObservacionesComponent } from './observaciones.component';
+import { MoleculaBaseComponent } from './molecula-base-.omponent';
+import { ValorNumericoComponent } from './valor-numerico.component';
+import { ValorNumericoFactory } from './valor-numerico.factory';
+import { ValorFechaComponent } from './valor-fecha.component';
+import { SolicitudPrestacionDefaultComponent } from './solicitud-default.component';
+import { SelectPorRefsetComponent } from './select-por-refset.component';
+import { RegistrarMedicamentoDefaultComponent } from './registrar-medicamento-default.component';
+import { LugarDeNacimientoComponent } from './lugar-nacimiento.componenet';
+import { LactanciaComponent } from './lactancia.component';
+import { FormulaBaseComponent } from './formula-base.component';
+import { SeleccionBinariaComponent } from './seleccion-binaria.component';
+import { EvolucionProblemaDefaultComponent } from './evolucion-problema-default.component';
+import { CalculoDeBostonComponent } from './calculo-boston.component';
+import { AutocitadoComponent } from './autocitado.component';
+import { AdjuntarDocumentoComponent } from './adjuntar-documento.component';
 import { SelectPrestacionComponent } from './select/select-prestacion.component';
 import { SelectOrganizacionComponent } from './select/select-organizacion.component';
 import { SelectSnomedComponent } from './select/select-snomed.component';
