+++ modules/descargas/informe-rup/elementos-rup/seleccion-binaria.component.ts
