+++ modules/descargas/informe-rup/elementos-rup/valor-numerico.factory.ts
