--- templates/rup/informes/html/includes/adjunto.html
@@ -1,8 +1,0 @@
-<div class="contenedor-concepto-data adjunto mb-4" style="page-break-before: always; page-break-after: always;">
-    <h3 class="text-uppercase">
-        Adjunto:
-        <!--descripcion-->
-    </h3>
-    <!--hr class="sm" /-->
-    <!--adjunto-->
-</div>
