--- templates/rup/informes/html/includes/adjuntoNoSoportado.html
@@ -1,10 +1,0 @@
-<article class="contenedor-data-origen adjunto">
-    <h3 class="text-uppercase">
-        Adjuntos no soportados:
-        <span class="text-sm">
-            Hay
-            <!--cantidadAdjuntos--> documentos adjuntos que no se pueden visualizar
-        </span>
-    </h3>
-</article>
-<br>
