--- templates/rup/informes/html/includes/hallazgo.html
@@ -1,10 +1,0 @@
-<article class="contenedor-concepto-data">
-    <p>
-        <!--concepto-->
-        <small>
-            <!--motivoPrincipalDeConsulta-->
-        </small>
-        <!--hr class="sm" /-->
-        <!--evolucion-->
-    </p>
-</article>
