--- templates/rup/informes/html/includes/insumo.html
@@ -1,28 +1,0 @@
-<article class="contenedor-concepto-data">
-    <p>
-        <!--concepto-->
-        <!--recetable-->
-
-        <small>
-            <!--motivoPrincipalDeConsulta-->
-        </small>
-        <!--hr class="sm" /-->
-        <span>
-            <b>Estado:</b>
-            <!--estado-->
-        </span>
-        <span>
-            <b>Cantidad:</b>
-            <!--cantidad-->
-            <!--unidad-->
-        </span>
-        <span>
-            <b>Durante:</b>
-            <!--cantidadDuracion-->
-            <!--uniadadDuracion-->
-        </span>
-    </p>
-    <p>
-        <!--indicacion-->
-    </p>
-</article>
