--- templates/rup/informes/html/includes/procedimiento.html
@@ -1,10 +1,0 @@
-<article class="contenedor-concepto-data">
-    <p>
-        <!--concepto-->
-        <small>
-            <!--motivoPrincipalDeConsulta-->
-        </small>
-        <!--hr class="sm" /-->
-        <!--valor-->
-    </p>
-</article>
