--- templates/rup/informes/html/includes/solicitud.html
@@ -1,23 +1,0 @@
-<article class="contenedor-concepto-data">
-    <h4>
-        Solicitud de:
-        <!--plan-->
-    </h4>
-    <!--hr class="sm" /-->
-    <p>
-        Motivo:
-        <!--motivo-->
-    </p>
-    <p>
-        Indicaciones:
-        <!--indicaciones-->
-    </p>
-    <p>
-        Organización:
-        <!--organizacionDestino-->
-    </p>
-    <p>
-        Profesional(es):
-        <!--profesionalesDestino-->
-    </p>
-</article>
