--- templates/rup/informes/html/informe.html
@@ -1,199 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-</head>
-
-<body>
-
-    <!-- Header -->
-    <header id="pageHeader">
-        <!-- Cabezal logos institucionales -->
-        <section class="contenedor-logos">
-            <span class="contenedor-logo-efector">
-                <!--logoOrganizacion-->
-            </span>
-            <span class="contenedor-logos-secundarios">
-                <!--logoAdicional-->
-                <!--logoAndes-->
-            </span>
-        </section>
-        <section class="contenedor-data-origen">
-            <!-- Datos paciente -->
-            <span class="contenedor-principal-data">
-                <div class="contenedor-secundario">
-                    <h6 class="volanta">Datos del paciente</h6>
-                    <h4>
-                        <!--paciente-->
-                    </h4>
-                    <h4>
-                        <!--datosRapidosPaciente-->
-                    </h4>
-                </div>
-                <div class="contenedor-secundario">
-                    <div class="contenedor-bloque-texto">
-                        <h6 class="bolder">Fecha de Nac.</h6>
-                        <h6>
-                            <!--fechaNacimiento-->
-                        </h6>
-                    </div>
-                    <div class="contenedor-bloque-texto">
-                        <h6 class="bolder">Nro. de carpeta</h6>
-                        <h6>
-                            <!--nroCarpeta-->
-                        </h6>
-                    </div>
-                    <div class="contenedor-bloque-texto">
-                        <h6>
-                            <!--obraSocial-->
-                        </h6>
-                    </div>
-                </div>
-            </span>
-
-            <!-- Datos origen solicitud -->
-            <span class="contenedor-principal-data">
-                <div class="contenedor-secundario">
-                    <h6 class="volanta">Datos de origen de solicitud</h6>
-                    <h4>
-                        <!--organizacionNombreSolicitud-->
-                    </h4>
-                    <h5>
-                        <!--organizacionDireccionSolicitud-->
-                    </h5>
-                </div>
-                <div class="contenedor-secundario">
-                    <div class="contenedor-bloque-texto">
-                        <h6 class="bolder">Profesional</h6>
-                        <h6>
-                            <!--profesionalSolicitud-->
-                        </h6>
-                    </div>
-                </div>
-            </span>
-        </section>
-        <h1 class="marca-de-agua">
-            <!--marca-de-agua-->
-        </h1>
-    </header>
-    <!-- END Header -->
-
-    <!-- Body -->
-    <main>
-        <section class="contenedor-informe">
-            <article class="cabezal-conceptos horizontal">
-                <!-- Titulo prestacion -->
-                <div class="contenedor-bloque-texto">
-                    <div class="tipo-prestacion">
-                        <!--tipoPrestacion-->
-                    </div>
-                </div>
-                <div class="contenedor-bloque-texto">
-                    <h6 class="bolder">
-                        <!--tituloFechaSolicitud-->
-                    </h6>
-                    <h6>
-                        <!--fechaSolicitud-->
-                    </h6>
-                </div>
-                <div class="contenedor-bloque-texto">
-                    <h6 class="bolder">
-                        <!--tituloFechaEjecucion-->
-                    </h6>
-                    <h6>
-                        <!--fechaEjecucion-->
-                    </h6>
-                </div>
-                <div class="contenedor-bloque-texto">
-                    <h6 class="bolder">
-                        <!--tituloFechaValidacion-->
-                    </h6>
-                    <h6>
-                        <!--fechaValidacion-->
-                    </h6>
-                </div>
-            </article>
-            <hr>
-            <div class="registros">
-                <!--registros-->
-            </div>
-        </section>
-    </main>
-
-    <!-- END Body -->
-
-    <!-- Footer -->
-    <footer id="last">
-        <!-- Firmas -->
-        <span class="contenedor-firmas">
-            <div class="contenedor-bloque-texto">
-                <!--firma1-->
-                <!-- <img src="assets/img/firmas/firma-1.png" alt=""> -->
-                <!--separadorFirma-->
-                <h6 class="bolder">
-                    <!--profesionalFirmante1-->
-                </h6>
-                <h6 class="subdata-italic"> </h6>
-                <h6 class="subdata-fecha">
-                    <!--fechaActual-->
-                </h6>
-            </div>
-        </span>
-
-    </footer>
-    <!-- Footer -->
-    <footer id="pageFooter">
-        <!-- Firmas -->
-        <span class="contenedor-firmas"></span>
-        <hr>
-        <span class="contenedor-zocalo">
-            <!--logoPDP-->
-            <!-- <img class="logo-pdp" src="assets/img/logo-pdp.png" alt=""> -->
-            <article class="contenedor-data-pdp">
-                <h6>Nota: El contenido de este informe ha sido validado digitalmente siguiendo los estándares de
-                    calidad y seguridad
-                    requeridos. El ministerio de salud de la provincia de Neuquén es responsable inscripto en el
-                    Registro
-                    Nacional de Protección de Datos Personales, según lo requiere la Ley N° 25.326 (art. 3° y 21 inciso
-                    1).</h6>
-            </article>
-            <article class="contenedor-data-organizacion">
-                <h6>
-                    <!--organizacionNombreSolicitud-->
-                </h6>
-                <h6>
-                    <!--organizacionDireccionSolicitud-->
-                </h6>
-            </article>
-            <article class="contenedor-data-impresion">
-                <h6 class="bolder">Impreso por:</h6>
-                <h6>
-                    <!--usuario-->
-                </h6>
-                <h6>
-                    <!--fechaActualFooter-->
-                </h6>
-            </article>
-            <article class="contenedor-data-validacion">
-                <h6 class="bolder">Validado por:</h6>
-                <h6>
-                    <!--profesionalValidacion-->
-                </h6>
-                <h6>
-                    <!--fechaValidacion-->
-                </h6>
-            </article>
-            <hr>
-            <span class="numeracion">
-                <small>{{page}}</small> de
-                <small>{{pages}}</small>
-            </span>
-        </span>
-    </footer>
-    <!-- END Footer -->
-
-</body>
-
-</html>
